Conflicts:
	ProjectSportyStuff/src/main/java/ourCode/Calculation.java

+++ ProjectSportyStuff/src/main/java/ourCode/Man.java
@@ -2,16 +2,16 @@ package ourCode;
 
 public class Man extends Athlete {
 
-	double onehundredM;
-	double longJump;
-	double shotPut;
-	double highJump;
-	double fourhundredM;
-	double onehundredtenMHurdles;
-	double discusThrow;
-	double poleVault;
-	double javelinThrow;
-	double onethousandfivehundredM;
+	int onehundredM;
+	int longJump;
+	int shotPut;
+	int highJump;
+	int fourhundredM;
+	int onehundredtenMHurdles;
+	int discusThrow;
+	int poleVault;
+	int javelinThrow;
+	int onethousandfivehundredM;
 
 	public Man(String firstName, String lastName, int id) {
 		this.firstName = firstName;
@@ -19,7 +19,7 @@ public class Man extends Athlete {
 		this.ID = id;
 	}
 
-	public double getOnehundredM() {
+	public int getOnehundredM() {
 		return onehundredM;
 	}
 
@@ -27,7 +27,7 @@ public class Man extends Athlete {
 		this.onehundredM = onehundredM;
 	}
 
-	public double getLongJump() {
+	public int getLongJump() {
 		return longJump;
 	}
 
@@ -35,7 +35,7 @@ public class Man extends Athlete {
 		this.longJump = longJump;
 	}
 
-	public double getShotPut() {
+	public int getShotPut() {
 		return shotPut;
 	}
 
@@ -43,7 +43,7 @@ public class Man extends Athlete {
 		this.shotPut = shotPut;
 	}
 
-	public double getHighJump() {
+	public int getHighJump() {
 		return highJump;
 	}
 
@@ -51,7 +51,7 @@ public class Man extends Athlete {
 		this.highJump = highJump;
 	}
 
-	public double getFourhundredM() {
+	public int getFourhundredM() {
 		return fourhundredM;
 	}
 
@@ -59,7 +59,7 @@ public class Man extends Athlete {
 		this.fourhundredM = fourhundredM;
 	}
 
-	public double getOnehundredtenMHurdles() {
+	public int getOnehundredtenMHurdles() {
 		return onehundredtenMHurdles;
 	}
 
@@ -67,7 +67,7 @@ public class Man extends Athlete {
 		this.onehundredtenMHurdles = onehundredtenMHurdles;
 	}
 
-	public double getDiscusThrow() {
+	public int getDiscusThrow() {
 		return discusThrow;
 	}
 
@@ -75,7 +75,7 @@ public class Man extends Athlete {
 		this.discusThrow = discusThrow;
 	}
 
-	public double getPoleVault() {
+	public int getPoleVault() {
 		return poleVault;
 	}
 
@@ -83,7 +83,7 @@ public class Man extends Athlete {
 		this.poleVault = poleVault;
 	}
 
-	public double getJavelinThrow() {
+	public int getJavelinThrow() {
 		return javelinThrow;
 	}
 
@@ -91,7 +91,7 @@ public class Man extends Athlete {
 		this.javelinThrow = javelinThrow;
 	}
 
-	public double getOnethousandfivehundredM() {
+	public int getOnethousandfivehundredM() {
 		return onethousandfivehundredM;
 	}
 
@@ -99,7 +99,7 @@ public class Man extends Athlete {
 		this.onethousandfivehundredM = onethousandfivehundredM;
 	}
 
-	public double getTotalResult() {
+	public int getTotalResult() {
 		return onehundredM + longJump + shotPut + highJump + fourhundredM + onehundredtenMHurdles + discusThrow
 				+ poleVault + javelinThrow + onethousandfivehundredM;
 	}
